gog: Use the urldb a fetch information about packages
urldb (first generated with the wiki2urldb script) is a dump of the db
stored in a wiki page:
  http://fedoraproject.org/wiki/Upstream_Release_Monitoring

This db is used to start populating our own db. It will need attention
to add and update packages from our obs instance.

The GNOME packages have a separate handling as we know a lot more about
those than any other random package. They follow strict rules that we
can use later on to be able to run the script with such intents as:
  'compare the latest stable|unstable versions of the packages we have
  in Trunk|devel:gnome'